tnc/requirements.txt
asgiref==3.8.1
astroid==3.3.8
asttokens==3.0.0
black==24.10.0
click==8.1.8
decorator==5.1.1
dill==0.3.9
django==5.1.4
django-debug-toolbar==5.0.1
django-extensions==3.2.3
executing==2.1.0
ipython==8.31.0
isort==5.13.2
jedi==0.19.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pip==24.3.1
platformdirs==4.3.6
prompt-toolkit==3.0.48
psycopg2==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.19.1
pylint==3.3.3
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
ruff==0.9.1
sqlparse==0.5.3
stack-data==0.6.3
tomlkit==0.13.2
traitlets==5.14.3
uv==0.5.18
wcwidth==0.2.13
